extract religiosity
