+++ lib/index.ts
@@ -74,7 +74,10 @@ export * from './npc-generation/createDescriptors'
 export * from './npc-generation/createLifestyleStandards'
 export * from './npc-generation/createName'
 export * from './npc-generation/createNamesake'
-export * from './npc-generation/createReligiosity'
+export * from './npc-generation/religiosity/createReligiosity'
+export * from './npc-generation/religiosity/getDeityProbabilities'
+export * from './npc-generation/religiosity/getReligiosityDescription'
+export * from './npc-generation/religiosity/pickDeity'
 export * from './npc-generation/createSocialClass'
 export * from './npc-generation/familyRaces'
 export * from './npc-generation/familyRelationships'
